+++ src/algorithms/mcpg/ising_default.yaml
@@ -2,7 +2,7 @@ problem_type: "ising"
 lr_init: 0.25
 regular_init: 0
 sample_epoch_num: 8
-max_epoch_num: 800
+max_epoch_num: 400
 reset_epoch_num: 80
 total_mcmc_num: 130
 repeat_times: 120
