+++ src/baseline/README.md
@@ -1,0 +1,11 @@
+
+# Baseline MIP Solvers
+
+
+
+| Solvers   | 
+|-----------|
+| Gurobi    |
+| ILOG CPLEX|
+| COPT      |
+
