+++ src/baseline/copt.py
@@ -1,39 +1,40 @@
 import networkx as nx
 import numpy as np
 import os
+import time
 import coptpy as cp
 from coptpy import COPT
 from utils import read_nxgraph
 from utils import float_to_binary
 from utils import base64_encode
 
-def save_to_file(model, file_path, time_limit, print_terminal=True):
+def save_to_file(model, file_path, time_limit, duration, print_terminal=True):
     
     file_dir = f"results/copt/{"/".join(file_path.split("/")[:-1])}"
     file_name = file_path.split("/")[-1]
 
     obj_val = model.objval
     obj_bnd = model.bestobj
-    # duration = model.runtime
-    # mip_gap = model.mipgap
-    best_solution = "".join([ float_to_binary(x.x) for x in model.getVars() ])
+    mip_gap = abs(obj_bnd - obj_val) / abs(obj_bnd)
+    best_solution = "".join([ float_to_binary(x.x) for x in model.getVars() if x.name[0] == "x" ])
     best_encoded = base64_encode(best_solution)
+
 
     with open(f"{file_dir}/{file_name}", "w") as f:
         f.write(f"Objective Value: {obj_val}\n")
         f.write(f"Objective Bound: {obj_bnd}\n")
-        # f.write(f"Duration: {duration}\n")
+        f.write(f"Duration: {duration}\n")
         f.write(f"Time Limit: {time_limit}\n")
-        # f.write(f"MIP Gap: {mip_gap}\n")
+        f.write(f"MIP Gap: {mip_gap}\n")
         f.write(f"Best Solution Encoded: {best_encoded}\n")
         f.write(f"Best Solution Raw: {best_solution}\n")
 
     if print_terminal:
         print(f"Objective Value: {obj_val}")
         print(f"Objective Bound: {obj_bnd}")
-        # print(f"Duration: {duration}")
+        print(f"Duration: {duration}")
         print(f"Time Limit: {time_limit}")
-        # print(f"MIP Gap: {mip_gap}")
+        print(f"MIP Gap: {mip_gap}")
         print(f"Best Solution Encoded: {best_encoded}")
         print(f"Best Solution Raw: {best_solution}\n")
 
@@ -67,8 +68,11 @@ def copt_solve(graph, file_path, time_limit=3600):
         os.makedirs(file_dir)
     model.setLogFile(f"{file_dir}/{file_name.split(".")[0]}.log")
 
+    start_time = time.time()
     model.solve()
-    save_to_file(model, file_path, time_limit)
+    end_time = time.time()
+    duration = (end_time - start_time)
+    save_to_file(model, file_path, duration, time_limit)
 
 def single_shot_instance(file_name, time_limit):
     graph = read_nxgraph(file_name)
@@ -80,7 +84,7 @@ def multiple_shot_instance(file_names, time_limit):
         single_shot_instance(file_name, time_limit)
 
 if __name__ == "__main__":
-    file_name = "data/RL/spin_glass/4x4/GSRL_16_ID0.txt"
+    file_name = "../../data/DRL/2D/10/DRL_10_ID0.txt"
     time_limit = 3600
     single_shot_instance(file_name, time_limit)
